edit at sweet alert in user management, create userform page.

+++ font_end/pages/User_Management.vue
@@ -4,19 +4,12 @@
     <v-row>
       <v-col cols="">
         <!-- Input for searching users by first name -->
-        <v-text-field
-          v-model="searchTerm"
-          label="Search"
-          prepend-icon="mdi-magnify"
-          single-line
-          hide-details
-        ></v-text-field>
+        <v-text-field v-model="searchTerm" label="Search" prepend-icon="mdi-magnify" single-line
+          hide-details></v-text-field>
       </v-col>
       <v-col cols="2" class="text-right">
         <!-- Button to add a new user -->
-        <v-btn color="primary" class="my-4 mx-4" @click="addUser"
-          >+ Add User</v-btn
-        >
+        <v-btn color="primary" class="my-4 mx-4" @click="addUser">+ Add User</v-btn>
       </v-col>
     </v-row>
 
@@ -24,25 +17,12 @@
 
     <v-row>
       <!-- Loop through users and display cards -->
-      <v-col
-        v-for="(user, index) in filteredUsers"
-        :key="index"
-        cols="12"
-        md="4"
-      >
+      <v-col v-for="(user, index) in filteredUsers" :key="index" cols="12" md="4">
         <!-- User card -->
         <v-card class="mx-auto" max-width="400">
           <!-- User image -->
-          <v-img
-            class="align-end text-white"
-            height="200"
-            :src="user.user_pic"
-            cover
-            @click="viewDetails(user)"
-          >
-            <v-card-title @click="viewDetails(user)"
-              >{{ user.user_firstname }} {{ user.user_lastname }}</v-card-title
-            >
+          <v-img class="align-end text-white" height="200" :src="user.user_pic" cover @click="viewDetails(user)">
+            <v-card-title @click="viewDetails(user)">{{ user.user_firstname }} {{ user.user_lastname }}</v-card-title>
           </v-img>
 
           <!-- User position -->
@@ -61,13 +41,9 @@
           <v-row class="mb-2">
             <v-col class="text-right" cols="12">
               <!-- Button to edit user -->
-              <v-btn class="mx-1" color="primary" @click="editUser(user)"
-                >Edit</v-btn
-              >
+              <v-btn class="mx-1" color="primary" @click="editUser(user)">Edit</v-btn>
               <!-- Button to delete user -->
-              <v-btn class="mx-4" color="primary" @click="deleteUser(user)"
-                >Delete</v-btn
-              >
+              <v-btn class="mx-4" color="primary" @click="deleteUser(user)">Delete</v-btn>
             </v-col>
           </v-row>
         </v-card>
@@ -81,41 +57,15 @@
         <v-card-text>
           <!-- Form to edit user details -->
           <v-form @submit.prevent="saveEditedUser">
-            <v-text-field
-              v-model="editedUser.user_firstname"
-              label="First Name"
-            ></v-text-field>
-            <v-text-field
-              v-model="editedUser.user_lastname"
-              label="Last Name"
-            ></v-text-field>
-            <v-select
-              v-model="editedUser.user_position"
-              :items="positions.map((position) => position.name)"
-              label="Position"
-            ></v-select>
-            <v-text-field
-              v-model="editedUser.user_department"
-              label="Department"
-            ></v-text-field>
-            <v-text-field
-              v-model="editedUser.user_email"
-              label="Email"
-            ></v-text-field>
-            <v-text-field
-              v-model="editedUser.user_password"
-              label="Password"
-            ></v-text-field>
-            <v-select
-              v-model="editedUser.user_status"
-              :items="statuses"
-              label="Status"
-            ></v-select>
-            <v-select
-              v-model="editedUser.user_role"
-              :items="roles"
-              label="Role"
-            ></v-select>
+            <v-text-field v-model="editedUser.user_firstname" label="First Name"></v-text-field>
+            <v-text-field v-model="editedUser.user_lastname" label="Last Name"></v-text-field>
+            <v-select v-model="editedUser.user_position" :items="positions.map((position) => position.name)"
+              label="Position"></v-select>
+            <v-text-field v-model="editedUser.user_department" label="Department"></v-text-field>
+            <v-text-field v-model="editedUser.user_email" label="Email"></v-text-field>
+            <v-text-field v-model="editedUser.user_password" label="Password"></v-text-field>
+            <v-select v-model="editedUser.user_status" :items="statuses" label="Status"></v-select>
+            <v-select v-model="editedUser.user_role" :items="roles" label="Role"></v-select>
             <label for="user_pic">Profile Picture:</label>
             <div>
               <!-- Input for selecting a new profile picture -->
@@ -149,6 +99,8 @@
 </template>
 
 <script>
+import Swal from 'sweetalert2';
+
 export default {
   methods: {
     // Navigate back to the User Management page
@@ -221,54 +173,45 @@ export default {
       }
     },
     // Save edited user details
+    // Save edited user details
     async saveEditedUser() {
-      // Check the structure of this.editedUser and this.image
-      console.log(this.editedUser);
-      console.log(this.image);
+      try {
+        // Check the structure of this.editedUser and this.image
+        console.log(this.editedUser);
+        console.log(this.image);
 
-      // Convert the image file to base64 format (if applicable)
-      const imageData = this.image
-        ? await this.getFileAsBase64(this.image)
-        : null;
+        // Convert the image file to base64 format (if applicable)
+        const imageData = this.image
+          ? await this.getFileAsBase64(this.image)
+          : null;
 
-      // Check if an image file is selected
-      if (imageData !== null) {
-        // Send a PUT request to the API to update the user with the new image
-        this.$axios
-          .put(`http://localhost:8080/api/users/${this.editedUser.user_id}`, {
+        // Check if an image file is selected
+        if (imageData !== null) {
+          // Send a PUT request to the API to update the user with the new image
+          await this.$axios.put(`http://localhost:8080/api/users/${this.editedUser.user_id}`, {
             ...this.editedUser,
             user_pic: imageData,
-          })
-          .then((response) => {
-            console.log("User updated successfully:", response.data);
-
-            // After updating, close the edit dialog
-            this.editDialog = false;
-
-            // After closing the edit dialog, refresh the user data by calling the API GET again
-            this.refreshUsersData();
-          })
-          .catch((error) => {
-            console.error("Error updating user:", error.response.data);
           });
-      } else {
-        // If no new image is selected, edit the user data without including the image
-        this.$axios
-          .put(`http://localhost:8080/api/users/${this.editedUser.user_id}`, {
+
+        } else {
+          // If no new image is selected, edit the user data without including the image
+          await this.$axios.put(`http://localhost:8080/api/users/${this.editedUser.user_id}`, {
             ...this.editedUser,
-          })
-          .then((response) => {
-            console.log("User updated successfully:", response.data);
-
-            // After updating, close the edit dialog
-            this.editDialog = false;
-
-            // After closing the edit dialog, refresh the user data by calling the API GET again
-            this.refreshUsersData();
-          })
-          .catch((error) => {
-            console.error("Error updating user:", error.response.data);
           });
+        }
+
+        // Display Sweet Alert for successful user update
+        Swal.fire('Success', 'User updated successfully.', 'success');
+
+        // After updating, close the edit dialog
+        this.editDialog = false;
+
+        // After closing the edit dialog, refresh the user data by calling the API GET again
+        await this.refreshUsersData();
+      } catch (error) {
+        // Handle any errors during the user update process
+        console.error('Error updating user:', error.response.data);
+        Swal.fire('Error', 'An error occurred during the user update process.', 'error');
       }
     },
     // Helper function to read a file and convert it to base64
@@ -308,15 +251,66 @@ export default {
             .startsWith(this.searchTerm.toLowerCase())
       );
     },
+    // Action to initiate user deletion with Sweet Alert
     // Action to initiate user deletion
     async deleteUser(user) {
-      this.userToDelete = user;
-      this.deleteDialog = true;
+      try {
+        // Display Sweet Alert for confirmation
+        const result = await Swal.fire({
+          title: 'Confirm Deletion',
+          text: `Are you sure you want to delete ${user.user_firstname} ${user.user_lastname}?`,
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#3085d6',
+          cancelButtonColor: '#d33',
+          confirmButtonText: 'Yes, delete it!',
+        });
+
+        if (result.isConfirmed) {
+          // If confirmed, proceed with the deletion
+          await this.deleteUserAction(user);
+          // Inform the user about the successful deletion
+          Swal.fire('Deleted!', 'User has been deleted.', 'success');
+        } else {
+          // If canceled, inform the user
+          Swal.fire('Cancelled', 'Deletion was cancelled', 'info');
+        }
+      } catch (error) {
+        // Handle any errors during the deletion process
+        console.error('Error initiating user deletion:', error);
+        Swal.fire('Error', 'An error occurred while initiating user deletion.', 'error');
+      }
     },
-    // Action to confirm user deletion
-    deleteConfirmed() {
-      this.deleteUserAction(this.userToDelete);
-      this.deleteDialog = false;
+    async deleteConfirmed() {
+      try {
+        // Display Sweet Alert for confirmation
+        const result = await Swal.fire({
+          title: 'Confirm Deletion',
+          text: `Are you sure you want to delete ${this.userToDelete.user_firstname} ${this.userToDelete.user_lastname}?`,
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#3085d6',
+          cancelButtonColor: '#d33',
+          confirmButtonText: 'Yes, delete it!',
+        });
+
+        if (result.isConfirmed) {
+          // If confirmed, proceed with the deletion
+          await this.deleteUserAction(this.userToDelete);
+          // Inform the user about the successful deletion
+          Swal.fire('Deleted!', 'User has been deleted.', 'success');
+        } else {
+          // If canceled, inform the user
+          Swal.fire('Cancelled', 'Deletion was cancelled', 'info');
+        }
+      } catch (error) {
+        // Handle any errors during the deletion process
+        console.error('Error deleting user:', error);
+        Swal.fire('Error', 'An error occurred during the deletion process.', 'error');
+      } finally {
+        // Close the delete dialog
+        this.deleteDialog = false;
+      }
     },
     // Action to cancel user deletion
     cancelDelete() {
